vulnscanner/requirements.txt
psutil
python-nmap
requests
rich 